atualização fontes
.

+++ src/main/java/com/br/gcm/controller/MunicipioController.java
@@ -6,8 +6,10 @@ import com.br.gcm.dao.UfDao;
 import com.br.gcm.model.MensagemTransacao;
 import com.br.gcm.model.Municipio;
 import com.br.gcm.model.Pais;
+import com.br.gcm.model.Usuario;
 import com.br.gcm.service.MunicipioService;
 import com.br.gcm.tag.Pagina;
+import com.br.gcm.util.Rotinas;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
@@ -36,13 +38,29 @@ public class MunicipioController {
     @Inject private PaisDao paisDao;
     @Inject private UfDao ufDao;
     @Inject private MunicipioService municipioService;
+    @Inject private Rotinas rotinas;
 
     @RequestMapping(value = "/municipio_lista")
     public String lista(@PageableDefault(size = 10) Pageable pageable, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1003");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100301");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100302");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100303");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100304");
+
         Municipio filtros = new Municipio();
         List<Pais> listaPais = paisDao.Pais_lista();
         filtros.setId_pais(listaPais.get(0).getId_pais());
 
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
         model.addAttribute("filtros", filtros);
         model.addAttribute("municipio_lista", MunicipioDao.selectAll(filtros, pageable));
         model.addAttribute("lista_pais", listaPais);
@@ -53,21 +71,40 @@ public class MunicipioController {
 
     //Filtros
     @RequestMapping(value = "/municipio_lista", method = RequestMethod.POST)
-    public ModelAndView filtros(@ModelAttribute Municipio filtros, @PageableDefault(size = 10) Pageable pageable) {
-        ModelAndView mav = new ModelAndView();
+    public String filtros(@ModelAttribute Municipio filtros, Model model, @PageableDefault(size = 10) Pageable pageable) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1003");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100301");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100302");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100303");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100304");
 
-        mav.addObject("lista_pais", paisDao.Pais_lista());
-        mav.addObject("municipio_lista", MunicipioDao.selectAll(filtros, pageable));
-        mav.addObject("pagina", new Pagina(pageable, MunicipioDao.count(filtros)));
-        mav.addObject("filtros", filtros);
-        mav.setViewName("municipio_lista");
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
+        model.addAttribute("lista_pais", paisDao.Pais_lista());
+        model.addAttribute("municipio_lista", MunicipioDao.selectAll(filtros, pageable));
+        model.addAttribute("pagina", new Pagina(pageable, MunicipioDao.count(filtros)));
+        model.addAttribute("filtros", filtros);
 
-        return mav;
+        return "municipio_lista";
     }
 
     //Deleta municipio
     @RequestMapping(value = "/municipio_deleta/{id_municipio}")
     public String deletar(@PathVariable("id_municipio") Integer id_municipio, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100303");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         try{
             municipioService.delete(id_municipio);
         }catch(Exception e){
@@ -80,6 +117,13 @@ public class MunicipioController {
     //novo
     @RequestMapping(value = "/municipio_novo", method = RequestMethod.GET)
     public String novo(ModelMap model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100301");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Municipio municipio = new Municipio();
         model.addAttribute("municipio", municipio);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
@@ -101,6 +145,13 @@ public class MunicipioController {
     //Editar municipio
     @RequestMapping(value = "/editar_municipio/{id}", method = RequestMethod.GET)
     public String edita(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100302");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Municipio municipio = MunicipioDao.selectById(id);
         model.addAttribute("municipio", municipio);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
@@ -123,6 +174,13 @@ public class MunicipioController {
     //Editar municipio
     @RequestMapping(value = "/detalhes_municipio/{id}", method = RequestMethod.GET)
     public String detalhes(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100304");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Municipio municipio = MunicipioDao.selectById(id);
         model.addAttribute("municipio", municipio);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
