+++ src/main/java/com/br/gcm/controller/PaisController.java
@@ -3,9 +3,11 @@ package com.br.gcm.controller;
 import com.br.gcm.dao.PaisDao;
 import com.br.gcm.model.MensagemTransacao;
 import com.br.gcm.model.Pais;
+import com.br.gcm.model.Usuario;
 import com.br.gcm.model.filtros.Filtro_Pais;
 import com.br.gcm.service.PaisService;
 import com.br.gcm.tag.Pagina;
+import com.br.gcm.util.Rotinas;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Controller;
@@ -26,11 +28,27 @@ import javax.swing.*;
 public class PaisController {
     @Inject private PaisDao paisDao;
     @Inject private PaisService paisService;
+    @Inject private Rotinas rotinas;
 
     @RequestMapping(value = "/pais_lista")
     public String pais_lista(@PageableDefault(size = 10) Pageable pageable, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1001");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100101");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100102");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100103");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100104");
+
         Filtro_Pais filtros = new Filtro_Pais();
 
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
         model.addAttribute("pais_lista", paisDao.Pais_Paginado(filtros, pageable));
         model.addAttribute("pagina", new Pagina(pageable, paisDao.count(filtros)));
         model.addAttribute("filtros", filtros);
@@ -40,20 +58,39 @@ public class PaisController {
 
     //Filtros
     @RequestMapping(value = "/pais_lista", method = RequestMethod.POST)
-    public ModelAndView filtros(@ModelAttribute Filtro_Pais filtros, @PageableDefault(size = 10) Pageable pageable) {
-        ModelAndView mav = new ModelAndView();
+    public String filtros(@ModelAttribute Filtro_Pais filtros, Model model, @PageableDefault(size = 10) Pageable pageable) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1001");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100101");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100102");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100103");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100104");
 
-        mav.addObject("pais_lista", paisDao.Pais_Paginado(filtros, pageable));
-        mav.addObject("pagina", new Pagina(pageable, paisDao.count(filtros)));
-        mav.addObject("filtros", filtros);
-        mav.setViewName("pais_lista");
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
+        model.addAttribute("pais_lista", paisDao.Pais_Paginado(filtros, pageable));
+        model.addAttribute("pagina", new Pagina(pageable, paisDao.count(filtros)));
+        model.addAttribute("filtros", filtros);
 
-        return mav;
+        return "pais_lista";
     }
 
     //Deleta Pais
     @RequestMapping(value = "/pais_deleta/{id_pais}")
     public String deletar(@PathVariable("id_pais") Integer id_pais, Model model) throws ServletException {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100103");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         try{
             paisService.delete(id_pais);
         }catch(Exception e){
@@ -66,6 +103,13 @@ public class PaisController {
     //Novo Pais
     @RequestMapping(value = "/pais_novo", method = RequestMethod.GET)
     public String novo(ModelMap model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100101");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Pais pais = new Pais();
         model.addAttribute("pais", pais);
         return "pais_novo";
@@ -85,6 +129,13 @@ public class PaisController {
 
     @RequestMapping(value = "/editar_pais/{id}", method = RequestMethod.GET)
     public String editar(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100102");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Pais pais = paisDao.selectById(id);
         model.addAttribute("pais", pais);
         return "pais_editar";
@@ -103,6 +154,13 @@ public class PaisController {
 
     @RequestMapping(value = "/pais_detalhes/{id}", method = RequestMethod.GET)
     public String detalhes(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100104");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Pais pais = paisDao.selectById(id);
         model.addAttribute("pais", pais);
         return "pais_detalhes";
